requirements.txt
stripe==5.7.0
python-dotenv==1.0.0
requests